// apps/web/src/lib/supabase/types.ts
export type products = {
  id: string;
  category_id: string;
  name: string;
  description: string | null;
  image_url: string | null;
  price: number;
  dietary_labels: string[] | null;
  available: boolean;
  sort_order: number;
  created_at: string;
  updated_at: string;
};